on VM

+++ string-array.cpp
@@ -30,7 +30,7 @@ void StringArray::addElem(GenericString* ptr){
 }
 
 
-String& StringArray::operator[](int index){
+String& StringArray::operator[](size_t index){
     if (index < array.size() && array[index] != nullptr) {
         return array[index]->as_string();
     } else {
@@ -41,7 +41,7 @@ String& StringArray::operator[](int index){
 
 }
 
-String& StringArray::operator[](int index) const{
+String& StringArray::operator[](size_t index) const{
 
     if (index < array.size() && array[index] != nullptr) {
         return array[index]->as_string();
